requirements.txt
asgiref==3.5.2
Django==4.1.2
Pillow==9.2.0
sqlparse==0.4.3
tzdata==2022.5
django-bootstrap5==22.1
djangorestframework==3.14.0
djoser==2.1.0